requirements.txt
Flask
llama-cpp-python
numpy
pyttsx3
pytz
requests
scipy
speechrecognition
sympy
# webrtcvad-wheels #sur windows
webrtcvad
whisper